+++ src/knowledge-memory/contracts/memory-lifecycle-profile.js
@@ -1,0 +1,59 @@
+export class MemoryLifecycleProfile {
+  constructor({
+    profileName,
+    objectives = [], memoryTypes = [], scopes = [], recordFields = [], candidateFields = [],
+    candidateSources = [], admissionDecisions = [], authorityClasses = [], sensitivityClasses = [],
+    lifecycleStates = [], transitionFields = [], writeOperations = [], readOperations = [],
+    degradedModes = [], evaluationDimensions = [], qualityAttributes = [], architecturalRules = [],
+    architectureBoundaries = [],
+    deliberateRetention = true, admissionRequired = true, candidatesSeparated = true,
+    modelOutputGoverned = true, conversationBoundary = true, executionStateBoundary = true,
+    domainAuthorityPreserved = true, policyAuthorityPreserved = true, memoryNonAuthoritative = true,
+    explicitRecordContract = true, scopePreserved = true, promotionGoverned = true,
+    materialChangeReadmitted = true, assertionStatesDistinct = true, temporalSemantics = true,
+    provenancePreserved = true, confidenceQualified = true, conflictVisible = true,
+    consolidationReadmitted = true, personalizationBounded = true, sensitiveInferenceControlled = true,
+    instructionBoundary = true, authoritiesSeparated = true, materialChangesVersioned = true,
+    lifecycleRecordAuthoritative = true, inactiveExcluded = true, uncertainResultsQualified = true,
+    completeLifecycle = true, correctionPropagated = true, deletionPropagated = true,
+    resurrectionPrevented = true, tombstonesMinimal = true, legalHoldSeparated = true,
+    retentionControlled = true, trainingSeparated = true, isolationEndToEnd = true,
+    lifecycleEvidenceProtected = true, idempotencyAndConcurrency = true, reconciliationEnabled = true,
+    safeDegradation = true, vendorNeutral = true, technologyIndependent = true,
+    transcriptCopiedByDefault = false, memoryGrantsAuthority = false, workflowStateStored = false,
+    agentStateStored = false, knowledgeAutoRemembered = false, modelDirectlyAdmits = false,
+    frequencyProvesTruth = false, scopeAutoPromoted = false, confidenceEqualsTruth = false,
+    contradictionsBlended = false, useExtendsRetention = false, derivedStoreAuthoritative = false,
+    legalHoldActivatesMemory = false, deletionLeavesRepresentations = false,
+    restoreReactivatesMemory = false, productionMemoryTrainsModels = false,
+    degradedModeWeakensControls = false, directDatabaseAccess = false, selectsMemoryProduct = false
+  }) {
+    this.profileName = profileName;
+    for (const [key, value] of Object.entries({
+      objectives, memoryTypes, scopes, recordFields, candidateFields, candidateSources,
+      admissionDecisions, authorityClasses, sensitivityClasses, lifecycleStates, transitionFields,
+      writeOperations, readOperations, degradedModes, evaluationDimensions, qualityAttributes,
+      architecturalRules, architectureBoundaries
+    })) this[key] = Object.freeze([...value]);
+    for (const [key, value] of Object.entries({
+      deliberateRetention, admissionRequired, candidatesSeparated, modelOutputGoverned,
+      conversationBoundary, executionStateBoundary, domainAuthorityPreserved,
+      policyAuthorityPreserved, memoryNonAuthoritative, explicitRecordContract, scopePreserved,
+      promotionGoverned, materialChangeReadmitted, assertionStatesDistinct, temporalSemantics,
+      provenancePreserved, confidenceQualified, conflictVisible, consolidationReadmitted,
+      personalizationBounded, sensitiveInferenceControlled, instructionBoundary,
+      authoritiesSeparated, materialChangesVersioned, lifecycleRecordAuthoritative,
+      inactiveExcluded, uncertainResultsQualified, completeLifecycle, correctionPropagated,
+      deletionPropagated, resurrectionPrevented, tombstonesMinimal, legalHoldSeparated,
+      retentionControlled, trainingSeparated, isolationEndToEnd, lifecycleEvidenceProtected,
+      idempotencyAndConcurrency, reconciliationEnabled, safeDegradation, vendorNeutral,
+      technologyIndependent, transcriptCopiedByDefault, memoryGrantsAuthority,
+      workflowStateStored, agentStateStored, knowledgeAutoRemembered, modelDirectlyAdmits,
+      frequencyProvesTruth, scopeAutoPromoted, confidenceEqualsTruth, contradictionsBlended,
+      useExtendsRetention, derivedStoreAuthoritative, legalHoldActivatesMemory,
+      deletionLeavesRepresentations, restoreReactivatesMemory, productionMemoryTrainsModels,
+      degradedModeWeakensControls, directDatabaseAccess, selectsMemoryProduct
+    })) this[key] = Boolean(value);
+    Object.freeze(this);
+  }
+}